moved encoding code to separate file
